refactor(javm-cap): move CapHashOrRef to cache.rs alongside CapRef
`CapHashOrRef` is the slot-target enum: `Hash(CapHash)` names a
`cache.blobs` entry, `Ref(CapRef)` is a `cache.instances` handle. Both
arms are cache-bound concepts, and the enum naturally lives next to
its two referents. Co-locates the directory / handle / slot triple
in one place.

`cap.rs` now `use super::cache::CapHashOrRef` for the
`Cap::CNode`-slot / `Cap::Instance.root_cnode` field references; same
for `cnode.rs`, `instance.rs`, `wire.rs`. The one external caller
that named the qualified path (`nub-arch-x86::call_loop`) is updated
to `javm_cap::cache::CapHashOrRef`. The canonical workspace path
`javm_cap::CapHashOrRef` is unchanged.
